// src/app/Models/Personal.ts
export interface IdDetails {
    name: string;
    idNo: number;
    issuePlace: string;
    issueDate: string;
    expiryDate: string;
  }
  
  export interface PersonalDetails {
    name: string;
    nationality: string;
    dateOfBirth: string;
    passportNo: number;
    profession: string;
  }
  